\name{apiFetch-package}
\alias{apiFetch-package}
\alias{apiFetch}
\docType{package}
\title{
\packageTitle{apiFetch}
}
\description{
\packageDescription{apiFetch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{apiFetch}
\packageIndices{apiFetch}

}
\author{
\packageAuthor{apiFetch}

Maintainer: \packageMaintainer{apiFetch}
}
\references{
See http://api.thenmap.net/doc/v2/ for more information on the thenmap API.
}

\keyword{ package }

\examples{
fetch()
fetch(date_str = "2020-01-01")
fetch(url = "http://api.thenmap.net/v2/fi-8/geo/")
}
