\name{apiFetch-package}
\alias{apiFetch-package}
\alias{apiFetch}
\docType{package}
\title{
\packageTitle{apiFetch}
}
\description{
\packageDescription{apiFetch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{apiFetch}
\packageIndices{apiFetch}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{apiFetch}

Maintainer: \packageMaintainer{apiFetch}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
