\name{apiFetch-package}
\alias{apiFetch-package}
\alias{apiFetch}
\docType{package}
\title{
\packageTitle{apiFetch}
}
\description{
\packageDescription{apiFetch}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{apiFetch}
\packageIndices{apiFetch}

}
\author{
\packageAuthor{apiFetch}

Maintainer: \packageMaintainer{apiFetch}
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
