% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{fetch}
\usage{
fetch(date_str = "2000-01-01", ...)
}
\arguments{
\item{date_str}{a date string expressed as "yyyy-mm-dd"}

\item{...}{optional parameters}
}
\value{
Returns a list with the data as well as a note if there was no data the specified date.
}
\description{
API fetch function
}
\examples{
fetch("2000-01-01")
}
